+++ src/compas_cra/equilibrium/__init__.py
@@ -1,6 +1,9 @@
 from .cra_pyomo import cra_solve
 from .cra_penalty_pyomo import cra_penalty_solve
 from .rbe_pyomo import rbe_solve
+from .rbe_robust import RobustForceResult
+from .rbe_robust import rbe_robust_sample
+from .rbe_robust import rbe_robust_support
 from .cra_helper import (
     equilibrium_setup,
     friction_setup,
@@ -27,6 +30,9 @@ __all__ = [
     "cra_solve",
     "cra_penalty_solve",
     "rbe_solve",
+    "rbe_robust_sample",
+    "rbe_robust_support",
+    "RobustForceResult",
     "equilibrium_setup",
     "friction_setup",
     "external_force_setup",